Break erb helper methods out into own file
Also add a few more so but due_ and opens_ have +- 3 days of helpers
